+++ apps/formio-server/src/middleware/pdfProxy/router.js
@@ -2,7 +2,6 @@
 
 const express = require('express');
 const fetch = require('node-fetch');
-const _ = require('lodash');
 const config = require('../../../config');
 const loadProjectContexts = require('../loadProjectContexts');
 const proxy = require('./proxy');
@@ -45,12 +44,7 @@ module.exports = (formioServer) => {
 
   // Keep only essential and user defined headers
   router.use((req, res, next) => {
-    const headers = {};
-    _.merge(headers,
-      _.pick(req.headers, 'accept', 'content-type', 'accept-encoding', 'accept-language'),
-      _.pickBy(req.headers, (_, h) => h.startsWith('x-'))
-    );
-    req.headers = headers;
+    proxy.updateHeadersForPdfRequest(req, formio);
     next();
   });