+++ apps/formio-server/src/models/models.js
@@ -3,6 +3,7 @@
 module.exports = function(router, models) {
   return {
     project: require('./Project')(router),
+    actionItem: require('./ActionItem')(router.formio),
     formrevision: require('./FormRevision')(router, models),
     submissionrevision: require('./SubmissionRevision')(router, models),
     session: require('./Session')(router),