+++ apps/formio-server/src/revisions/Revision.js
@@ -1,7 +1,7 @@
 'use strict';
 const _ = require('lodash');
 const async = require('async');
-
+const config = require('../../config');
 module.exports = class Revision {
   constructor(app, type, trackedProperties, revisionPlans) {
     this.revisionPlans = revisionPlans || ['trial', 'commercial'];
@@ -12,15 +12,25 @@ module.exports = class Revision {
     this.idToBson = app.formio.formio.util.idToBson;
   }
 
-  checkRevisionPlane(plan) {
+  checkRevisionPlan(plan) {
     return this.revisionPlans.includes(plan);
   }
 
+  revisionsAllowed(req) {
+    return this.checkRevisionPlan(req.primaryProject.plan) &&
+      !config.formio.hosted &&
+      (this.app.license && !this.app.license.licenseServerError && _.get(req, 'licenseTerms.options.sac', false));
+  }
+
   shouldCreateNewRevision(req, item, loadItem, form) {
+    if (!this.revisionsAllowed(req)) {
+      return false;
+    }
+
     const currentFormTrackedProperties = _.pick(loadItem, this.trackedProperties);
     const updatedFormTrackedProperties = _.pick(req.body, this.trackedProperties);
     const isChanged = !_.isEqual(currentFormTrackedProperties, updatedFormTrackedProperties);
-    const areRevisionsAllowed = (item.revisions || form && form[`${this.type}Revisions`]) && this.revisionPlans.includes(req.primaryProject.plan);
+    const areRevisionsAllowed = (item.revisions || form && form[`${this.type}Revisions`]);
     return (
       req.isDraft ||
       item.revisions && !form.revisions ||