+++ apps/formio-server/test/tests/tags.js
@@ -750,6 +750,20 @@ module.exports = function(app, template, hook) {
     describe('Deployments', () => {
       let env1, env2, form, formWithEnabledRevisions, formRevisions, parentForm, resource, resourceWithEnabledRevisions, resourceRevisions, parentRecourse, action, role, tag, _export;
 
+      before(done => {
+        if (!config.formio.hosted) {
+          process.env.TEST_SIMULATE_SAC_PACKAGE = '1';
+        }
+        done();
+      });
+  
+      after(done => {
+        if (!config.formio.hosted) {
+          process.env.TEST_SIMULATE_SAC_PACKAGE = '0';
+        }
+        done();
+      });
+
       it('Create Environment 1', done => {
         const myProject = {
           title: chance.word(),